Se modularizo el home creando un componente para renderizarlo en el index de pages.
